requirements.txt
Flask
google-genai
requests
python-dotenv
beautifulsoup4
lxml
cachetools